+++ lib/spreedly.rb
@@ -23,6 +23,7 @@ require 'spreedly/transactions/retain_payment_method'
 require 'spreedly/transactions/redact_payment_method'
 require 'spreedly/transactions/redact_gateway'
 require 'spreedly/gateway'
+require 'spreedly/gateway_class'
 require 'spreedly/error'
 
 module Spreedly