+++ lib/spreedly/environment.rb
@@ -87,6 +87,11 @@ module Spreedly
       Gateway.new_list_from(xml_doc)
     end
 
+    def gateway_options
+      xml_doc = ssl_options(gateway_options_url)
+      GatewayClass.new_list_from(xml_doc)
+    end
+
     def add_gateway(gateway_type, credentials = {})
       body = add_gateway_body(gateway_type, credentials)
       xml_doc = ssl_post(add_gateway_url, body, headers)